fix a bug in previous commit

+++ le-pls/src/main/webapp/app/AppCommon/services/TopPredictorService.js
@@ -559,7 +559,7 @@ angular.module('mainApp.appCommon.services.TopPredictorService', [
     };
 
         this.createTicks = function(maxTickValue, maxTickNumber) {
-            if (typeof maxTickValue === 'string') { maxTickValue = parseInt(maxTickValue); }
+            if (typeof maxTickValue === 'string') { maxTickValue = parseFloat(maxTickValue); }
             var steps = [0.5, 1, 2, 5, 10];
             // iterate options in steps, find the maximum appropriate step
             var step = _.reduce(steps, function(memo, s){
